ApaPlan_OJT/requirement.txt
dash
flask
firebase-admin
requests
python-dotenv
flask-login
dash-mantine-components
PyJWT
requests
gunicorn
